server/python_service/requirements.txt
flask
pdfplumber
python-docx
scikit-learn
